Фото на память (выделил валидацию в модуль)
